create propagation_functions to offload code off instrumentSimu_functions
