+++ Asterix/Example_param_file.ini
@@ -15,7 +15,7 @@ Data_dir = './'
     wavelength_0 = 780e-9  
 
     # Spectral band (in meters) I think should be in percent...
-    Delta_wav = 0#10e-9
+    Delta_wav = 50e-9
 
     # Number of monochromatic images in the spectral band. 
     # Ignored if Delta_wav = 0
@@ -23,7 +23,7 @@ Data_dir = './'
     nb_wav = 3
 
     #Dimension of science image
-    dim_im   = 400  #400        # image size (in pixels)
+    dim_im   = 600  #400        # image size (in pixels)
 
     #Sampling in the science image
     #(lambda/Lyot diameter) in pixel
